1.3.1
- Checks and doesn't allow for for 0's in Jewellery tools and user settings
- Added ability to regenerate API key from UserAccounts
-- Added a little tick next to the button that shows when you've regenerated that API key (disappears on page reload)
- Removed ability to turn off your own admin status in UserAccounts (shows error popup)
- Added a Hovering Text wrapper for buttons / other elements
-- Added to most small buttons eg. EditableTable Delete/Add, UserAccounts Delete/Switch Admin Status/Regenerate API Key etc
- Added extra handling for null values in APIs in both Frontend and Backend
- Fixed error where output box for RollingWire page showed "g" instead of "mm"

+++ Frontier.Server/Controllers/ErrorsController.cs
@@ -31,14 +31,17 @@ namespace Frontier.Server.Controllers
         [HttpDelete("{base64ErrorId}/delete")]
         public async Task<IActionResult> DeleteErrorLog(string base64ErrorId)
         {
-            string errorId = functions.ConvertFromBase64(base64ErrorId);
+            if (base64ErrorId != null) {
+                string errorId = functions.ConvertFromBase64(base64ErrorId);
 
-            // Check if the error exists
-            ErrorLedgerModel error = await db.GetErrorLog(errorId);
-            if (error == null) return NotFound("Error not found");
+                // Check if the error exists
+                ErrorLedgerModel error = await db.GetErrorLog(errorId);
+                if (error == null) return NotFound("Error not found");
 
-            await db.DeleteErrorLog(errorId);
-            return Ok();
+                await db.DeleteErrorLog(errorId);
+                return Ok();
+            }
+            return BadRequest("No Error Id Supplied");
         }
     }
 }
